+++ digiwf-integrations/digiwf-dms-integration/digiwf-dms-integration-core/src/test/java/de/muenchen/oss/digiwf/dms/integration/adapter/in/streaming/DepositObjectStreamingAdapterTest.java
@@ -1,6 +1,5 @@
-package de.muenchen.oss.digiwf.dms.integration.adapter.in;
+package de.muenchen.oss.digiwf.dms.integration.adapter.in.streaming;
 
-import de.muenchen.oss.digiwf.dms.integration.adapter.in.streaming.CancelObjectDto;
 import de.muenchen.oss.digiwf.message.process.api.error.IncidentError;
 import jakarta.validation.ValidationException;
 import org.junit.jupiter.api.Assertions;
@@ -18,24 +17,24 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-class CancelObjectStreamingAdapterTest extends MessageProcessorTestBase {
+class DepositObjectStreamingAdapterTest extends MessageProcessorTestBase {
 
-    private final CancelObjectDto cancelObjectDto = new CancelObjectDto(
+    private final DepositObjectDto depositObjectDto = new DepositObjectDto(
             "objectCoo",
             "user"
     );
-    private Message<CancelObjectDto> message;
+    private Message<DepositObjectDto> message;
 
     @BeforeEach
     void setup() {
         setupBase();
-        Mockito.doNothing().when(cancelObjectInPortMock).cancelObject(
-                cancelObjectDto.getObjectCoo(),
-                cancelObjectDto.getUser());
+        Mockito.doNothing().when(depositObjectInPortMock).depositObject(
+                depositObjectDto.getObjectCoo(),
+                depositObjectDto.getUser());
         this.message = new Message<>() {
             @Override
-            public CancelObjectDto getPayload() {
-                return cancelObjectDto;
+            public DepositObjectDto getPayload() {
+                return depositObjectDto;
             }
 
             @Override
@@ -46,15 +45,15 @@ class CancelObjectStreamingAdapterTest extends MessageProcessorTestBase {
     }
 
     @Test
-    void testCancelObjectSuccessful() {
-        streamingAdapter.cancelObject().accept(this.message);
-        verify(cancelObjectInPortMock, times(1)).cancelObject(cancelObjectDto.getObjectCoo(), cancelObjectDto.getUser());
+    void testDepositObjectSuccessful() {
+        streamingAdapter.depositObject().accept(this.message);
+        verify(depositObjectInPortMock, times(1)).depositObject(depositObjectDto.getObjectCoo(), depositObjectDto.getUser());
     }
 
     @Test
-    void testCancelObjectValidationException() {
-        Mockito.doThrow(new ValidationException("Test ValidationException")).when(cancelObjectInPortMock).cancelObject(any(), any());
-        streamingAdapter.cancelObject().accept(this.message);
+    void testDepositObjectValidationException() {
+        Mockito.doThrow(new ValidationException("Test ValidationException")).when(depositObjectInPortMock).depositObject(any(), any());
+        streamingAdapter.depositObject().accept(this.message);
         final ArgumentCaptor<Map> messageHeaderArgumentCaptor = ArgumentCaptor.forClass(Map.class);
         verify(errorApiMock, times(1)).handleIncident(messageHeaderArgumentCaptor.capture(), any(IncidentError.class));
         Assertions.assertTrue(messageHeaderArgumentCaptor.getValue().containsKey(DIGIWF_PROCESS_INSTANCE_ID));
@@ -62,9 +61,9 @@ class CancelObjectStreamingAdapterTest extends MessageProcessorTestBase {
 
 
     @Test
-    void testCancelObjectIncidentError() {
-        Mockito.doThrow(new IncidentError("Error Message")).when(cancelObjectInPortMock).cancelObject(any(), any());
-        streamingAdapter.cancelObject().accept(this.message);
+    void testDepositObjectIncidentError() {
+        Mockito.doThrow(new IncidentError("Error Message")).when(depositObjectInPortMock).depositObject(any(), any());
+        streamingAdapter.depositObject().accept(this.message);
         final ArgumentCaptor<Map> messageHeaderArgumentCaptor = ArgumentCaptor.forClass(Map.class);
         verify(errorApiMock, times(1)).handleIncident(messageHeaderArgumentCaptor.capture(), any(IncidentError.class));
         Assertions.assertTrue(messageHeaderArgumentCaptor.getValue().containsKey(DIGIWF_PROCESS_INSTANCE_ID));
